+++ src/misc/progress/interfaces.ts
@@ -20,14 +20,17 @@ export interface IProgressConstructor {
 export interface IProgressJSON extends Writeable<Pick<IProgress, 'loaded' | 'total' | 'lengthComputable' | 'name'>> {
 }
 
+
+
+
 /**
  * A Progress is like a ProgressEvent, but doesn't implement all unnecessary properties
  */
-export interface IProgress {
+export interface IProgress extends IProgressOptions {
   readonly lengthComputable: boolean; // returns true if the length is computable
   readonly loaded: number; // number of bytes, percents, etc... loaded
-  readonly total: number; // total number of bytes, percents, etc... to load. INFO may be Number.POSITIVE_INFINITE, if length if not computable (lengthComputable => false)
-  readonly name?: string;
+  readonly total: number; // total number of bytes, percents, etc... to load. INFO MUST be Number.POSITIVE_INFINITE, if length if not computable (lengthComputable => false)
+  readonly name?: string; // name may be used to do the distinction between different kind of progress
 
   toJSON(allowFloat?: boolean): IProgressJSON; // because JSON doesnt support 'Infinite', use lengthComputable.
   toString(): string;
