+++ src/misc/progress/public.ts
@@ -1,2 +1,3 @@
 export * from './interfaces';
 export { Progress, IsProgress } from './implementation';
+export * from './functions';
