+++ src/notifications/observables/task/context/constructor.ts
@@ -1,0 +1,19 @@
+import { ITaskContext } from './interfaces';
+import { ITask } from '../interfaces';
+import { ConstructClassWithPrivateMembers } from '../../../../misc/helpers/ClassWithPrivateMembers';
+import { ITaskContextInternal, TASK_CONTEXT_PRIVATE } from './privates';
+
+let ALLOW_TASK_CONTEXT_CONSTRUCT: boolean = false;
+
+export function AllowTaskContextConstruct(allow: boolean): void {
+  ALLOW_TASK_CONTEXT_CONSTRUCT = allow;
+}
+
+export function ConstructTaskContext<TValue>(instance: ITaskContext<TValue>, task: ITask<TValue>): void {
+  if (ALLOW_TASK_CONTEXT_CONSTRUCT) {
+    ConstructClassWithPrivateMembers(instance, TASK_CONTEXT_PRIVATE);
+    (instance as ITaskContextInternal<TValue>)[TASK_CONTEXT_PRIVATE].task = task;
+  } else {
+    throw new TypeError('Illegal constructor');
+  }
+}
